+++ front/src/pages/DeviceParts.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Select, Table, Icon, Button, message, Modal } from "antd";
+import { Select, Table, Icon, Button, message, Modal, Breadcrumb } from "antd";
 import ApiUtil from "../Utils/ApiUtil";
 import HttpUtil from "../Utils/HttpUtil";
 import DevicePartDialog from "./DevicePartDialog";
@@ -16,7 +16,7 @@ export default class DeviceParts extends React.Component {
     },
     {
       title: "产品名称",
-      dataIndex: "name"
+      render: item => <span style={{ cursor: "pointer" }} onClick={() => this.onItemClick(item)}>{item.name}</span>
     },
     {
       title: "单位",
@@ -49,36 +49,42 @@ export default class DeviceParts extends React.Component {
     mItems: [],
     showInfoDialog: false,
     currentItem: {},
-    mPartTypes: [{ id: 0, name: "配件大类" }, ...CommonValues.partTypes]
+    mTypes: [{ id: 0, name: "配件大类" }, ...CommonValues.partTypes]
   };
 
   currentType = 0;
 
   getData(type) {
-    if (type === 0) {
-      this.setState({
-        mItems: CommonValues.partTypes
-      });
-    } else {
-      HttpUtil.get(ApiUtil.API_GET_PARTS + type)
+    HttpUtil.get(ApiUtil.API_GET_PARTS + type)
       .then(data => {
         data.map((item, index) => (item.index = index + 1));
-        this.setState({mItems: data});
+        if (type === 0) {   // 大类
+          this.setState({
+            mTypes: [{ id: 0, name: '配件大类' }, ...data],
+            mItems: data
+          });
+        } else {
+          this.setState({
+            mItems: data
+          });
+        }
       })
       .catch(error => {
         message.error(error.message);
       });
-    }
-    
   }
 
   removeData(id) {
     HttpUtil.get(ApiUtil.API_PART_DELETE + id)
       .then(
         re => {
+          let code = re.code;
+          if (code < 0) {
+            throw new Error(re.message)
+          }
           message.info(re.message);
           let items = ArrayUtil.deleteItem(this.state.mItems, id);
-          this.setState({mItems: items});
+          this.setState({ mItems: items });
         }
       ).catch(error => {
         message.error(error.message);
@@ -90,12 +96,13 @@ export default class DeviceParts extends React.Component {
   }
 
 
-  renderDialog(){
+  renderDialog() {
     if (this.state.showInfoDialog) {
       return (
         <DevicePartDialog
           visible={this.state.showInfoDialog}
           part={this.state.currentItem}
+          itemTypes={this.state.mTypes}
           onClose={() => this.setState({ showInfoDialog: false })}
           onDialogConfirm={this.onDialogConfirm}
         />
@@ -103,31 +110,54 @@ export default class DeviceParts extends React.Component {
     }
   }
 
-  renderColumns(){
-    return this.currentType === 0 ? [...this.columns].splice(0, 2) : this.columns;
+  getTypeName(id) {
+    for (let i = 0; i < this.state.mTypes.length; i++) {
+      let item = this.state.mTypes[i];
+      if (item.id === id) {
+        return item.name;
+        break;
+      }
+    }
   }
+
+  renderBreadcrumb() {
+    let style = { display: 'inline-block', fontSize: '16px', cursor: "pointer", color: '#1890ff' };
+    if (this.currentType == 0) {
+      return (<Breadcrumb style={style}>
+        <Breadcrumb.Item>配件大类</Breadcrumb.Item></Breadcrumb>);
+    } else {
+      return (<Breadcrumb style={style}>
+        <Breadcrumb.Item onClick={() => this.handleFilterChange(0)}>
+          <Icon type='left' style={{ marginRight: '6px' }} />
+          配件大类
+        </Breadcrumb.Item>
+        <Breadcrumb.Item>{this.getTypeName(this.currentType)}</Breadcrumb.Item>
+      </Breadcrumb>);
+    }
+  }
+
+  renderColumns() {
+    if (this.currentType === 0) {
+      let array = [...this.columns];
+      array.splice(2, 3);
+      return array;
+    } else {
+      return this.columns;
+    }
+  }
+
 
   render() {
     return (
       <div>
-        <div>
-          <Select
-            style={{ width: 240, marginRight: 20, marginTop: 4 }}
-            defaultValue={this.currentType}
-            onChange={this.handleFilterChange}
-          >
-            {this.state.mPartTypes.map(item => (
-              <Select.Option value={item.id} key={item.id + ""}>
-                {item.name}
-              </Select.Option>
-            ))}
-          </Select>
+        <div style={{ display: 'flex', paddingTop: 10, alignItems: 'center' }}>
+          {this.renderBreadcrumb()}
 
           <Button
             type="primary"
             icon="plus"
             onClick={() => this.showUpdateDialog()}
-            style={{ float: "right", marginTop: 4 }}
+            style={{ marginLeft: 'auto' }}
           >
             添加
           </Button>
@@ -137,7 +167,6 @@ export default class DeviceParts extends React.Component {
           dataSource={this.state.mItems}
           rowKey={item => item.id}
           columns={this.renderColumns()}
-          onRow={this.currentType === 0 && this.onClickRow}
           pagination={false}
         />
 
@@ -148,15 +177,13 @@ export default class DeviceParts extends React.Component {
   }
 
   // 点击行
-  /* onClickRow = item => {
-    return {
-      onClick: () => {
-        //message.info(item.name);
-        this.currentType = item.id;
-        this.getData(this.currentType);
-      }
-    };
-  }; */
+  onItemClick = item => {
+    if (this.currentType == 0) {
+      this.handleFilterChange(item.id);
+    } else {
+      this.showUpdateDialog(item);
+    }
+  }
 
 
   onDialogConfirm = (part, newId) => {
@@ -171,13 +198,25 @@ export default class DeviceParts extends React.Component {
         mItems: datas
       });
     } else {    // 新增
-      part.id = newId;
-      part.index = this.state.mItems.length+1;
-      let datas = [...this.state.mItems];
-      datas.push(part);
-      this.setState({
-        mItems: datas
-      });
+      // 如果不在当前类别下就刷新，不然就在当前页动态添加。
+      if (part.sid != this.currentType) {
+        this.handleFilterChange(part.sid)
+      } else {
+        part.id = newId;
+        part.index = this.state.mItems.length + 1;
+        let datas = [...this.state.mItems];
+        datas.push(part);
+        this.setState({
+          mItems: datas
+        });
+        if (part.sid === 0) {
+          let types = [...this.state.mTypes];
+          types.push(part);
+          this.setState({
+            mTypes: types
+          })
+        }
+      }
     }
   }
 
@@ -192,7 +231,7 @@ export default class DeviceParts extends React.Component {
       currentItem = {
         id: 0,
         name: '',
-        sType: this.currentType
+        sid: this.currentType
       };
     }
     this.setState({
